+++ fbw-common/src/systems/instruments/src/ND/ND.tsx
@@ -700,13 +700,17 @@ class TopMessages extends DisplayComponent<{
 
   private readonly trueTrackWord = Arinc429RegisterSubject.createEmpty();
 
-  private needApprMessageUpdate = true;
+  private readonly apprMessage0 = ConsumerSubject.create(this.sub.on('apprMessage0'), 0);
 
-  private apprMessage0: number;
+  private readonly apprMessage1 = ConsumerSubject.create(this.sub.on('apprMessage1'), 0);
 
-  private apprMessage1: number;
-
-  private readonly approachMessageValue = Subject.create('');
+  private readonly approachMessageValue = MappedSubject.create(
+    ([apprmsg0, apprmsg1]) => {
+      return SimVarString.unpack([apprmsg0, apprmsg1]);
+    },
+    this.apprMessage0,
+    this.apprMessage1,
+  );
 
   private readonly btvMessageValue = Subject.create('');
 
@@ -757,29 +761,11 @@ class TopMessages extends DisplayComponent<{
     super.onAfterRender(node);
 
     this.sub
-      .on('apprMessage0')
-      .whenChanged()
-      .handle((value) => {
-        this.apprMessage0 = value;
-        this.needApprMessageUpdate = true;
-      });
-
-    this.sub
-      .on('apprMessage1')
-      .whenChanged()
-      .handle((value) => {
-        this.apprMessage1 = value;
-        this.needApprMessageUpdate = true;
-      });
-
-    this.sub
       .on('ndBtvMessage')
       .whenChanged()
       .handle((value) => {
         this.btvMessageValue.set(value);
       });
-
-    this.sub.on('simTime').whenChangedBy(100).handle(this.refreshApproachMessage.bind(this));
 
     this.sub.on('trueTrackRaw').handle((v) => this.trueTrackWord.setWord(v));
 
@@ -791,15 +777,6 @@ class TopMessages extends DisplayComponent<{
       .on('trueRefActive')
       .whenChanged()
       .handle((v) => this.trueRefActive.set(!!v));
-  }
-
-  private refreshApproachMessage(): void {
-    if (this.needApprMessageUpdate) {
-      const ident = SimVarString.unpack([this.apprMessage0, this.apprMessage1]);
-
-      this.approachMessageValue.set(ident);
-      this.needApprMessageUpdate = false;
-    }
   }
 
   render(): VNode | null {